+++ packages/transcription-whispercpp/config.cmake
@@ -1,3 +1,0 @@
-# Allows model addon to find required symbols
-set(MLC_HIDE_PRIVATE_SYMBOLS OFF)
-